+++ src/frontend/sider/Fagsak/Behandling/Sider/Vedtak/Vedtaksbrev.tsx
@@ -7,7 +7,7 @@ import { RessursStatus } from '@navikt/familie-typer';
 import FeilutbetaltValuta from './FeilutbetaltValuta/FeilutbetaltValuta';
 import RefusjonEøs from './RefusjonEøs/RefusjonEøs';
 import SammensattKontrollsak from './SammensattKontrollsak/SammensattKontrollsak';
-import { useSammensattKontrollsak } from './SammensattKontrollsak/useSammensattKontrollsak';
+import { useSammensattKontrollsak } from './SammensattKontrollsak/SammensattKontrollsakContext';
 import { BehandlingKorrigertAlert } from './Vedtak';
 import { VedtaksbegrunnelseTeksterProvider } from './VedtakBegrunnelserTabell/Context/VedtaksbegrunnelseTeksterContext';
 import VedtaksperioderMedBegrunnelser from './VedtakBegrunnelserTabell/VedtaksperioderMedBegrunnelser/VedtaksperioderMedBegrunnelser';
