+++ src/frontend/sider/Fagsak/Behandling/Sider/Vedtak/Vedtaksmeny.tsx
@@ -14,7 +14,7 @@ import { ASpacing10 } from '@navikt/ds-tokens/dist/tokens';
 
 import KorrigerEtterbetaling from './KorrigerEtterbetaling/KorrigerEtterbetaling';
 import KorrigerVedtak from './KorrigerVedtakModal/KorrigerVedtak';
-import { useSammensattKontrollsak } from './SammensattKontrollsak/useSammensattKontrollsak';
+import { useSammensattKontrollsak } from './SammensattKontrollsak/SammensattKontrollsakContext';
 import EndreEndringstidspunkt from './VedtakBegrunnelserTabell/endringstidspunkt/EndreEndringstidspunkt';
 import { Behandlingstype, type IBehandling } from '../../../../../typer/behandling';
 import { BehandlingKategori } from '../../../../../typer/behandlingstema';
